requirements.txt
# Python API
fastapi
loguru
uvicorn
locust
pycocotools
zipp
ultralytics
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-prometheus
opentelemetry-exporter-jaeger
python-dotenv
fastapi
pydantic
uvicorn
python-decouple
python-multipart
pyflakes
numpy
# Requirements for models
torch
